refactor(views): Update form-step10 and form-buttons blade templates (feature/da_minor_issues)
The form-step10.blade.php template has been modified to remove the
:last-selected-friend-id prop from the StudentFadeComponent. This prop is
no longer needed.

In the form-buttons.blade.php template, the button with the name "back-btn"
has been added for steps 12 and 14. The button with the name "back-btn" has
been modified to dispatch the 'set-sub-step-id-down' event for steps 12 and 14.
The button with the name "back-btn" has been modified to dispatch the
'set-sub-step-id-up' event for steps other than 12 and 14. The button with
the type "submit" has been modified to disable the button when loading.

The student-fade-component.blade.php template has been modified to add the
class "selected-btn-shrink" to the button when the $